requirements.txt
fastapi
uvicorn
aiogram
sqlalchemy
psycopg2-binary
requests
python-dotenv
pydantic